# schemify: inline non-authentic struct predicate