--- a/racket/src/schemify/schemify.rkt
+++ b/racket/src/schemify/schemify.rkt
@@ -838,17 +838,28 @@
                                     prim-knowns knowns imports mutated simples unsafe-mode?)]
                 [else #f]))
             (define (inline-struct-predicate k s-rator im args)
-              (define type-id (and (known-struct-predicate-authentic? k)
-                                   (pair? args)
+              (define type-id (and (pair? args)
                                    (null? (cdr args))
                                    (inline-type-id k im add-import! mutated imports)))
               (cond
-                [type-id
+                [(not type-id) #f]
+                [(known-struct-predicate-authentic? k)
                  (define tmp (maybe-tmp (car args) 'v))
                  (define ques `(unsafe-struct? ,tmp ,(schemify type-id 'fresh)))
                  (wrap-tmp tmp (car args)
                            ques)]
-                [else #f]))
+                [else
+                 (define tmp (maybe-tmp (car args) 'v))
+                 (define schemified-type-id (schemify type-id 'fresh))
+                 (define tmp-type-id (maybe-tmp schemified-type-id 'v))
+                 (define ques `(if (unsafe-struct? ,tmp ,tmp-type-id)
+                                   #t
+                                   (if (impersonator? ,tmp)
+                                       (unsafe-struct? (impersonator-val ,tmp) ,tmp-type-id)
+                                       #f)))
+                 (wrap-tmp tmp (car args)
+                           (wrap-tmp tmp-type-id schemified-type-id 
+                                     ques))]))
             (define (inline-field-access k s-rator im args)
               ;; Inline the selector with an `unsafe-struct?` test plus `unsafe-struct*-ref`.
               (define type-id (and (pair? args)