Currently non-authentic struct predicates are not inlined, and the overhead seems quite significant.
For example, on my machine the following program
```racket
#lang racket
(struct A ())

(define (f x)
  (for/fold ([a #t])
            ([_ (in-range 100000000)])
    (A? x)))
(define a #f)
(set! a (A))

(collect-garbage)
(time (f a))
```
costs
380ms current
200ms patched
140ms authentic