@mflatt 2021-01-05
I've wondered about whether to do this. How much does code size increase with the change?