@yjqww6 2021-01-05
Code size from `./racket` -> `(collect-garbage)` -> `(dump-memory-stats)` (ta6osx)
master
```
code                      15,371   13,455,664  |     15,376   13,456,528
```
patched
```
 code                      15,374   13,536,352  |     15,379   13,537,216
```
I would try DrRacket later. If the code size indeed grows too much, probably https://github.com/yjqww6/racket/commit/0ef382bd68265c1c49762c89b74b6139a34b5b72 is better.