@mflatt 2021-01-05
That other approach does look better, in any case. For what it's worth, I'd move the `(known-struct-predicate-authentic? k)` around the use of `tmp` instead of mostly duplicating the clause.