@yjqww6 2021-01-06
Code sizes from `./racket` -> `(require drracket)` (and waiting for the window) -> `(collect-garbage)` -> `(dump-memory-stats)`, rebuild using `./raco setup -c` and `./raco setup -D`
master
```
 code                     152,249  130,140,832  |    152,254  130,141,696
 total                  7,309,850  506,834,492  |  7,477,242  515,688,380
```
patched
```
 code                     151,986  130,247,712  |    151,991  130,248,576
 total                  7,285,189  506,000,292  |  7,452,573  514,845,860
```
https://github.com/yjqww6/racket/commit/0ef382bd68265c1c49762c89b74b6139a34b5b72
```
 code                     151,971  130,125,136  |    151,976  130,126,000
 total                  7,284,634  505,877,148  |  7,452,018  514,722,716
```
Though this might not be an ideal approach and varies across each runs, code size seems not to be a problem.

benchmark (in another machine):
250ms master
110ms patched
110ms authentic
200ms https://github.com/yjqww6/racket/commit/0ef382bd68265c1c49762c89b74b6139a34b5b72

contracted case:
```racket
#lang racket
(struct A (a) #:mutable)

(define (f x)
  (for/fold ([a #t])
            ([_ (in-range 100000000)])
    (A? x)))

(define/contract a
  (struct/c A integer?) (A 0))

(collect-garbage)
(time (f a))
```
450ms master
330ms patched
220ms https://github.com/yjqww6/racket/commit/0ef382bd68265c1c49762c89b74b6139a34b5b72

Another program:
```racket
#lang racket
(require racket/unsafe/ops)

(struct A (a b c d))
(define (build n)
  (cond
    [(> n 0)
     (define a (build (sub1 n)))
     (A a a a a)]
    [else 1]))
(define (f x)
  (match x
    [(A a b c d)
     (unsafe-fx+
      (f a)
      (f b)
      (f c)
      (f d))]
    [else 1]))

(define t (build 13))
(collect-garbage)
(time (f t))
```
600ms master
310ms patched
250ms authentic
450ms https://github.com/yjqww6/racket/commit/0ef382bd68265c1c49762c89b74b6139a34b5b72