@mflatt 2021-01-06
Looks like this is the best option - thanks!